# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Tapture backend
  version: '1'
  description:
    Browsers may call the API cross-origin only from the exact origins in CORS_ORIGINS. Those origins
    receive Access-Control-Allow-Origin and Vary Origin, and preflights answer 204 allowing GET, POST, PATCH
    and DELETE with the Authorization, Content-Type, X-Api-Version and Idempotency-Key headers. Other
    preflights answer 403. Credentials are never allowed. An empty CORS_ORIGINS sends no CORS headers.
paths:
...
  /api/v1/projects:
    get:
      responses:
        '200':
          description: Visible projects.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Project'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
        '413':
          $ref: '#/components/responses/Error'
        '429':
          $ref: '#/components/responses/Error'
        '500':
          $ref: '#/components/responses/Error'
        '503':
          $ref: '#/components/responses/Error'
      parameters: []
    post:
      responses:
        '201':
          description: Registered.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
        '413':
          $ref: '#/components/responses/Error'
        '429':
          $ref: '#/components/responses/Error'
        '500':
          $ref: '#/components/responses/Error'
        '503':
          $ref: '#/components/responses/Error'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id: *a1
                name: *a1
              required:
                - id
                - name
              additionalProperties: false
      parameters: []
  /api/v1/projects/{id}:
    patch:
      responses:
        '200':
          description: Name or relay settings.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
        '413':
          $ref: '#/components/responses/Error'
        '429':
          $ref: '#/components/responses/Error'
        '500':
          $ref: '#/components/responses/Error'
        '503':
          $ref: '#/components/responses/Error'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: *a1
                relayEnabled: &a3
                  type: boolean
                neverRelay: *a3
                retentionDays: &a7
                  type: integer
                  minimum: 1
                  maximum: 90
              required: []
              additionalProperties: false
      parameters:
        - name: id
          in: path
          required: true
          schema: *a1
  /api/v1/projects/{id}/members:
    get:
      responses:
        '200':
          description: Members.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Membership'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
        '413':
          $ref: '#/components/responses/Error'
        '429':
          $ref: '#/components/responses/Error'
        '500':
          $ref: '#/components/responses/Error'
        '503':
          $ref: '#/components/responses/Error'
      parameters:
        - name: id
          in: path
          required: true
          schema: *a1
    post:
      responses:
        '201':
          description: Added.
          content:
            application/json:
              schema:
                type: object
                properties:
                  userId: *a1
                required:
                  - userId
                additionalProperties: false
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
        '413':
          $ref: '#/components/responses/Error'
        '429':
          $ref: '#/components/responses/Error'
        '500':
          $ref: '#/components/responses/Error'
        '503':
          $ref: '#/components/responses/Error'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userId: *a1
                contextScope: &a6
                  type: string
                  nullable: true
              required:
                - userId
              additionalProperties: false
      parameters:
        - name: id
          in: path
          required: true
          schema: *a1
  /api/v1/projects/{id}/members/{userId}:
    delete:
      responses:
        '204':
          description: Removed.
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
        '413':
          $ref: '#/components/responses/Error'
        '429':
          $ref: '#/components/responses/Error'
        '500':
          $ref: '#/components/responses/Error'
        '503':
          $ref: '#/components/responses/Error'
      parameters:
        - name: id
          in: path
          required: true
          schema: *a1
        - name: userId
          in: path
          required: true
          schema: *a1
  /api/v1/projects/{id}/relay/packages:
    get:
      responses:
        '200':
          description: Waiting packages.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PackagePage'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
        '413':
          $ref: '#/components/responses/Error'
        '429':
          $ref: '#/components/responses/Error'
        '500':
          $ref: '#/components/responses/Error'
        '503':
          $ref: '#/components/responses/Error'
      parameters:
        - name: id
          in: path
          required: true
          schema: *a1
        - name: cursor
          in: query
          schema: *a1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
    post:
      responses:
        '201':
          description: Stored opaque package.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Package'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
        '413':
          $ref: '#/components/responses/Error'
        '429':
          $ref: '#/components/responses/Error'
        '500':
          $ref: '#/components/responses/Error'
        '503':
          $ref: '#/components/responses/Error'
      parameters:
        - name: id
          in: path
          required: true
          schema: *a1
      requestBody:
        required: true
        content:
          application/octet-stream:
            schema:
              type: string
              format: binary
  /api/v1/projects/{id}/relay/packages/{packageId}:
    get:
      responses:
        '200':
          description: Opaque bytes.
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
        '413':
          $ref: '#/components/responses/Error'
        '429':
          $ref: '#/components/responses/Error'
        '500':
          $ref: '#/components/responses/Error'
        '503':
          $ref: '#/components/responses/Error'
      parameters:
        - name: id
          in: path
          required: true
          schema: *a1
        - name: packageId
          in: path
          required: true
          schema: *a1
...
components:
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              enum:
                - invalid_request
                - unauthorized
                - invalid_credentials
                - forbidden
                - not_found
                - conflict
                - payload_too_large
                - rate_limited
                - quota_exceeded
                - internal
                - unavailable
            message: *a1
            details:
              type: object
              additionalProperties: true
          required:
            - code
            - message
          additionalProperties: false
      required:
        - error
      additionalProperties: false
...
    Project:
      type: object
      properties:
        id: *a1
        organisationId: *a1
        name: *a1
        relayEnabled: *a3
        neverRelay: *a3
        retentionDays: *a7
      required:
        - id
        - organisationId
        - name
        - relayEnabled
        - neverRelay
        - retentionDays
      additionalProperties: false
    Membership:
      type: object
      properties:
        projectId: *a1
        userId: *a1
        contextScope: *a6
      required:
        - projectId
        - userId
        - contextScope
      additionalProperties: false
    Package:
      type: object
      properties:
        id: *a1
        projectId: *a1
        authorDeviceId: *a1
        byteSize: *a4
        createdAt: *a1
        expiresAt: *a1
        storageRef: *a1
      required:
        - id
        - projectId
        - authorDeviceId
        - byteSize
        - createdAt
        - expiresAt
        - storageRef
      additionalProperties: false
    PackagePage:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Package'
        nextCursor: *a6
      required:
        - items
        - nextCursor
      additionalProperties: false
...
  responses:
    Error:
      description: Typed safe error envelope.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'